QRCode/requirements.txt
streamlit==1.41.1
qrcode[pil]==8.0
Pillow==11.0.0
